+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to the package will be documented here.
 
 ---
 
+## [4.26.61] - 2025-11-17
+
+- **Fixed** issue with `Chronos` *format methods* not formatting correctly when `useUTC` is `true`.
+- **Updated** *tsdoc* for some `Chronos` methods.
+
 ## [4.26.60] - 2025-11-16
 
 - **Fixed** an *issue* in `Chronos` class where chaining *calculation methods* after `timeZone`, `toLocal`, `toUTC`, or other *offset-adjusting methods* resulted in *incorrect* time values.